test: move ts tests in a separate workspace, add test for ts v4, collect coverage only once
