+++ package.json
@@ -105,10 +105,12 @@
     "format": "prettier \"{,**/}*.{ts,tsx,mts,js,json,md}\" --check",
     "format:fix": "prettier \"{,**/}*.{ts,tsx,mts,js,json,md}\" --write",
     "pretest": "npm run generate_ts_v4_index",
+    "pretest:typescript": "npm run generate_ts_v4_index",
     "test": "vitest --run",
-    "test:coverage": "vitest --project runtime --project compatibility --coverage --run",
+    "test:coverage": "vitest --coverage --run",
     "test:runtime": "vitest --project runtime",
     "test:compatibility": "vitest --project compatibility",
+    "test:typescript": "vitest --workspace vitest.workspace.typescript.mts",
     "test:local": "vitest --workspace vitest.workspace.local.mts",
     "build": "rimraf dist && rollup -c && echo '{\"type\":\"module\"}' > dist/esm/package.json && cpy \"./dist/umd/*.js\" ./",
     "generate_ts_v4_index": "cp index.d.ts index.v4.d.ts && node -e \"fs.writeFileSync('index.v4.d.ts', fs.readFileSync('index.v4.d.ts').toString().replace(/t.js/g, 't.v4.js').replace(/export type \\* /g, '// export type * '))\"",
